@@ -1,20 +1,26 @@
 {
   "name": "generator-trigen-frontend",
-  "version": "2.0.0-alpha",
+  "version": "2.1.0-alpha",
   "description": "Yeoman generator for frontend development.",
-  "main": "generators/index.js",
-  "repository": "https://github.com/TrigenSoftware/generator-trigen-frontend",
   "author": "dangreen",
   "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/TrigenSoftware/generator-trigen-frontend.git"
+  },
+  "bugs": {
+    "url": "https://github.com/TrigenSoftware/generator-trigen-frontend/issues"
+  },
+  "main": "generators/app/index.js",
   "engines": {
     "node": ">=6.4.0"
   },
   "scripts": {
-    "eslint": "eslint --cache 'generators/app/*.js' 'generators/app/prompts/*.js' 'test/*/{,src/,tasks/}*.js'"
+    "lint": "eslint --cache 'src/app/*.js' 'src/app/prompts/*.js' 'test/*/{,src/,tasks/}*.js'",
+    "build": "rollup -c",
+    "watch": "rollup -c -w",
+    "prepublishOnly": "npm run build"
   },
-  "files": [
-    "generators"
-  ],
   "keywords": [
     "trigen",
     "frontend",
@@ -25,13 +31,45 @@
     "yeoman-generator"
   ],
   "dependencies": {
-    "chalk": "^1.1.3",
+    "chalk": "^2.3.0",
     "hasbin": "^1.2.3",
-    "yeoman-generator": "^1.0.0",
-    "yosay": "^1.2.1"
+    "yeoman-generator": "^2.0.2",
+    "yosay": "^2.0.1"
   },
   "devDependencies": {
-    "eslint": "^3.16.1",
-    "eslint-config-trigen": "^1.2.0"
-  }
+    "babel-core": "^6.26.0",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "escope": "^3.6.0",
+    "eslint": "^4.15.0",
+    "eslint-config-trigen": "^2.3.1",
+    "rollup": "^0.54.0",
+    "rollup-plugin-babel": "^3.0.3",
+    "rollup-plugin-commonjs": "^8.2.6",
+    "rollup-plugin-eslint": "^4.0.0",
+    "rollup-plugin-json": "^2.3.0",
+    "rollup-plugin-node-resolve": "^3.0.2"
+  },
+  "babel": {
+    "presets": [
+      [
+        "env",
+        {
+          "targets": {
+            "node": "6"
+          }
+        }
+      ],
+      "stage-0"
+    ],
+    "plugins": [
+      "add-module-exports",
+      "transform-runtime"
+    ]
+  },
+  "files": [
+    "generators"
+  ]
 }
