@@ -1,14 +1,14 @@
 
-module.exports =
-function managePackageDeps(sourcePkg, props) {
+export default function managePackageDeps(sourcePkg, props) {
 
 	const { dependencies, devDependencies } = sourcePkg;
 
-	return Object.assign({}, sourcePkg, {
+	return {
+		...sourcePkg,
 		dependencies:    applyProps(dependencies, props),
 		devDependencies: applyProps(devDependencies, props)
-	});
-};
+	};
+}
 
 function applyProps(deps, props) {
 	return Object.keys(deps).reduce((result, dep) => {
@@ -20,12 +20,18 @@ function applyProps(deps, props) {
 				propVal = props[propKey];
 
 			if (not && !propVal || !not && propVal) {
-				return Object.assign({}, result, deps[dep]);
+				return {
+					...result,
+					...deps[dep]
+				};
 			}
 
 			return result;
 		}
 
-		return Object.assign({}, result, { [dep]: deps[dep] });
+		return {
+			...result,
+			[dep]: deps[dep]
+		};
 	}, {});
 }
